core: Pull serialization out of request execution
This commit moves the construction of the payload out of the main
execution request flow into it's own function. This is working towards
the goal of allowing individual API operations to hook serialization,
but fall back to the conventional defaults if no hook is provided.

Also allow separate encoding from structure - this allows a hook to just
return a struct with appropriate tags for JSON encoding while retaining
a default.