test: Make tests use abdullin/seq library
This removes some (monstrous) homegrown reflection in favour of using
Rinat's Seq library for asserting against responses in acceptance tests.
This adds support for nested structures using "Object path" notation
(e.g. "StorageAccount.PrimaryEndpoint.Queue" == "helloworld", and
"ARecords[0].ipv4Address" == "10.1.1.1").

The only oddity right now is that if types have JSON tagging on them,
the JSON tag names must be used in the assertion paths and not the field
names (technically it is always using the JSON tags, but the default
naming means that by default these match the field names in the struct).