vm: Add CreateOrUpdateVirtualMachine operation
So far this is untested - it will need many acceptance tests to verify
operation. The documentation for this is inconsistent, incomplete, and
probably wrong in multiple places. For example:

Field Name: certificateUrl
Required: yes
Type: String
Description:
    It is the Base64 encoding of the following JSON Object
    which is encoded in UTF-8:

    ```
    {
      "data":"<Base64-encoded-certificate>",
      "dataType":"pfx",
      "password":"<pfx-file-password>"
    }
    ```

The example request shows an actual URL, however. Riviera takes a
string, so to a point it is up to the caller to resolve which section of
the documentation to believe.

Several other components are either missing from the documentation or
documented only via the example request - I have taken guesses at what
form they may take.